Refactoring. Split endpoints into separate mixins.
